LayoutTests/editing/selection/containsNode-with-no-common-ancestor.html

<!DOCTYPE html>
<html>
<body>
<input type="text" value="hello">
<script src="../../resources/js-test.js"></script>
<script>

description(`This tests calling containsNode when the selection is set inside a shadow tree.<br>
WebKit should not hit a debug assertion`);

shouldBeFalse(`document.querySelector('input').select(); getSelection().containsNode(document.body)`);

</script>
</body>
</html>
